Split actions from reducers
